requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
cached-property==1.5.2
    # via kombu
celery==5.1.2
    # via
    #   -r requirements/base.in
    #   django-celery-results
certifi==2021.5.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.4
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
decorator==5.0.9
    # via ipython
django==3.2.6
    # via
    #   -r requirements/base.in
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   django-redis
    #   djangorestframework
django-celery-results==2.2.0
    # via -r requirements/base.in
django-environ==0.4.5
    # via -r requirements/base.in
django-extensions==3.1.3
    # via -r requirements/base.in
django-filter==2.4.0
    # via -r requirements/base.in
django-model-utils==4.1.1
    # via -r requirements/base.in
django-redis==5.0.0
    # via -r requirements/base.in
djangorestframework==3.12.4
    # via -r requirements/base.in
flake8==3.9.2
    # via
    #   -r requirements/lint.in
    #   flake8-pep3101
    #   flake8-polyfill
    #   flake8-print
    #   flake8-quotes
    #   pep8-naming
flake8-import-order==0.18.1
    # via -r requirements/lint.in
flake8-pep3101==1.3.0
    # via -r requirements/lint.in
flake8-polyfill==1.0.2
    # via pep8-naming
flake8-print==4.0.0
    # via -r requirements/lint.in
flake8-quotes==3.2.0
    # via -r requirements/lint.in
idna==3.2
    # via requests
importlib-metadata==4.6.4
    # via
    #   flake8
    #   kombu
    #   pep517
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
ipython==7.26.0
    # via -r requirements/base.in
ipython-genutils==0.2.0
    # via traitlets
jedi==0.18.0
    # via ipython
kombu==5.1.0
    # via celery
matplotlib-inline==0.1.2
    # via ipython
mccabe==0.6.1
    # via flake8
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pep517==0.11.0
    # via pip-tools
pep8-naming==0.12.1
    # via -r requirements/lint.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.3.1
    # via -r requirements/base.in
pip-tools==6.2.0
    # via -r requirements/base.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.20
    # via
    #   click-repl
    #   ipython
psycopg2==2.9.1
    # via -r requirements/base.in
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via
    #   flake8
    #   flake8-import-order
    #   flake8-print
pyflakes==2.3.1
    # via flake8
https://github.com/kradalby/pyGE/archive/master/0.0.4.zip
    # via -r requirements/base.in
pygments==2.10.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   -r requirements/base.in
    #   pytest-pythonpath
pytest-pythonpath==0.7.3
    # via -r requirements/base.in
pytz==2021.1
    # via
    #   celery
    #   django
redis==3.5.3
    # via django-redis
requests==2.26.0
    # via pyge
sentry-sdk==1.3.1
    # via -r requirements/base.in
six==1.16.0
    # via
    #   click-repl
    #   flake8-print
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via pytest
tomli==1.2.1
    # via pep517
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   importlib-metadata
urllib3==1.26.6
    # via
    #   requests
    #   sentry-sdk
uwsgi==2.0.19.1
    # via -r requirements/base.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.1
    # via -r requirements/base.in
wheel==0.37.0
    # via pip-tools
zipp==3.5.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
